+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Fix `let_it_be` issue when initialized with an array/enumerable or an AR relation. ([@pirj][])
+- Improve `RSpec/AggregateExamples` (formerly `RSpec/AggregateFailures`) cop. ([@pirj][])
 
 ## 0.10.2 (2020-01-07) 🎄